requirements.txt
python-dotenv
pandas
numpy
sqlalchemy